Updating Clique to Elemental revision 759:bdf9ce4a332b
--HG--
rename : external/elemental/examples/special_matrices/Cauchy.cpp => external/elemental/examples/matrices/Cauchy.cpp
rename : external/elemental/examples/special_matrices/CauchyLike.cpp => external/elemental/examples/matrices/CauchyLike.cpp
rename : external/elemental/examples/special_matrices/Circulant.cpp => external/elemental/examples/matrices/Circulant.cpp
rename : external/elemental/examples/special_matrices/Diagonal.cpp => external/elemental/examples/matrices/Diagonal.cpp
rename : external/elemental/examples/special_matrices/DiscreteFourier.cpp => external/elemental/examples/matrices/DiscreteFourier.cpp
rename : external/elemental/examples/special_matrices/Hankel.cpp => external/elemental/examples/matrices/Hankel.cpp
rename : external/elemental/examples/special_matrices/HermitianUniformSpectrum.cpp => external/elemental/examples/matrices/HermitianUniformSpectrum.cpp
rename : external/elemental/examples/special_matrices/Hilbert.cpp => external/elemental/examples/matrices/Hilbert.cpp
rename : external/elemental/examples/special_matrices/Identity.cpp => external/elemental/examples/matrices/Identity.cpp
rename : external/elemental/examples/special_matrices/Legendre.cpp => external/elemental/examples/matrices/Legendre.cpp
rename : external/elemental/examples/special_matrices/NormalUniformSpectrum.cpp => external/elemental/examples/matrices/NormalUniformSpectrum.cpp
rename : external/elemental/examples/special_matrices/OneTwoOne.cpp => external/elemental/examples/matrices/OneTwoOne.cpp
rename : external/elemental/examples/special_matrices/Ones.cpp => external/elemental/examples/matrices/Ones.cpp
rename : external/elemental/examples/special_matrices/Toeplitz.cpp => external/elemental/examples/matrices/Toeplitz.cpp
rename : external/elemental/examples/special_matrices/Uniform.cpp => external/elemental/examples/matrices/Uniform.cpp
rename : external/elemental/examples/special_matrices/Walsh.cpp => external/elemental/examples/matrices/Walsh.cpp
rename : external/elemental/examples/special_matrices/Wilkinson.cpp => external/elemental/examples/matrices/Wilkinson.cpp
rename : external/elemental/examples/special_matrices/Zeros.cpp => external/elemental/examples/matrices/Zeros.cpp

+++ external/elemental/tests/blas-like/Gemm.cpp
@@ -6,7 +6,9 @@
    which can be found in the LICENSE file in the root directory, or at 
    http://opensource.org/licenses/BSD-2-Clause
 */
-#include "elemental.hpp"
+#include "elemental-lite.hpp"
+#include "elemental/blas-like/level3/Gemm.hpp"
+#include "elemental/matrices/Uniform.hpp"
 using namespace std;
 using namespace elem;
 
@@ -226,8 +228,7 @@ main( int argc, char* argv[] )
                  << "Testing with doubles:\n"
                  << "---------------------" << endl;
         }
-        TestGemm<double>
-        ( print, orientA, orientB, m, n, k, (double)3, (double)4, g );
+        TestGemm<double>( print, orientA, orientB, m, n, k, 3., 4., g );
 
         if( commRank == 0 )
         {
