+++ external/elemental/tests/blas-like/Symm.cpp
@@ -6,7 +6,9 @@
    which can be found in the LICENSE file in the root directory, or at 
    http://opensource.org/licenses/BSD-2-Clause
 */
-#include "elemental.hpp"
+#include "elemental-lite.hpp"
+#include "elemental/blas-like/level3/Symm.hpp"
+#include "elemental/matrices/Uniform.hpp"
 using namespace std;
 using namespace elem;
 
@@ -109,7 +111,7 @@ main( int argc, char* argv[] )
                  << "Testing with doubles:\n"
                  << "---------------------" << endl;
         }
-        TestSymm<double>( side, uplo, m, n, (double)3, (double)4, print, g ); 
+        TestSymm<double>( side, uplo, m, n, 3., 4., print, g ); 
 
         if( commRank == 0 )
         {
