+++ external/elemental/tests/blas-like/Syrk.cpp
@@ -6,7 +6,9 @@
    which can be found in the LICENSE file in the root directory, or at 
    http://opensource.org/licenses/BSD-2-Clause
 */
-#include "elemental.hpp"
+#include "elemental-lite.hpp"
+#include "elemental/blas-like/level3/Syrk.hpp"
+#include "elemental/matrices/Uniform.hpp"
 using namespace std;
 using namespace elem;
 
@@ -107,8 +109,7 @@ main( int argc, char* argv[] )
                  << "Testing with doubles:\n"
                  << "---------------------" << endl;
         }
-        TestSyrk<double>
-        ( print, uplo, orientation, m, k, (double)3, (double)4, g );
+        TestSyrk<double>( print, uplo, orientation, m, k, 3., 4., g );
 
         if( commRank == 0 )
         {
